agents/research_c/pyproject.toml
[project]
name = "research-c-agent"
version = "0.1.0"
description = "Research Agent C microservice"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk==1.22.*",
    "a2a-sdk==0.3.*",
    "fastapi==0.123.*",
    "uvicorn==0.40.0",
    "google-cloud-logging==3.13.0",
    "opentelemetry-exporter-gcp-trace==1.11.0",
    "httpx==0.28.*",
    "python-dotenv==1.0.*",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
